backend/pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.20.0",
    "fastapi>=0.115.0",
    "openai-agents>=0.6.2",
    "python-dotenv>=1.0.0",
    "qdrant-client>=1.16.1",
    "requests>=2.32.5",
    "trafilatura>=2.0.0",
    "uvicorn>=0.32.0",
]
